Fixed errors and moved labels
